+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Misti 0.4.1 support
 
+### Fixed
+- Removed filepath and MistiResult hacks
+
 ## [0.4.0] - 2024-10-09
 
 ### Added