+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Misti 0.5.0 support
+
 ## [0.4.1] - 2024-10-12
 
 ### Added